pyproject.toml
[project]
name = "image-diff"
version = "0.1.0"
description = ""
authors = [
    {name = "Patrick",email = "patrsc@nubu.at"}
]
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "pillow (>=11.3.0,<12.0.0)",
    "imagehash (>=4.3.2,<5.0.0)",
    "torch (>=2.2.0,<2.3.0)",
    "imagededup (>=0.3.3.post2,<0.4.0)",
    "numpy (<2)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
